Make cached endpoint seperate
